pyproject.toml
[project]
name = "regelrecht-mvp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12.0"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.6",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "ruff>=0.14.4",
]
